+++ lib/chat_web/live/main_live/index.ex
@@ -80,6 +80,13 @@ defmodule ChatWeb.MainLive.Index do
     |> noreply()
   end
 
+  @impl true
+  def handle_params(%{"hash" => hash}, _, %{assigns: %{live_action: :chat_link}} = socket) do
+    socket
+    |> assign(:chat_link, hash)
+    |> noreply()
+  end
+
   def handle_params(_, _, socket), do: socket |> noreply()
 
   @impl true
@@ -334,6 +341,34 @@ defmodule ChatWeb.MainLive.Index do
   def handle_event("cargo:activate", _params, socket) do
     UsbDriveDumpRoom.remove()
     CargoRoom.activate(socket.assigns.room.pub_key)
+
+    %CargoSettings{} = cargo_settings = socket.assigns.cargo_settings
+    %Identity{} = me = socket.assigns.me
+    %Identity{} = room_identity = socket.assigns.room_identity
+    %Room{} = room = socket.assigns.room
+
+    cargo_settings.checkpoints
+    |> Enum.each(fn checkpoint_pub_key ->
+      dialog = Dialogs.open(me, checkpoint_pub_key |> User.by_id())
+
+      invites =
+        Dialogs.read(dialog, me)
+        |> Stream.filter(fn %{type: type} -> type == :room_invite end)
+        |> Stream.filter(fn message ->
+          identity = Dialogs.extract_invite_room_identity(message)
+          identity.public_key == room_identity.public_key
+        end)
+        |> Enum.to_list()
+
+      if invites == [] do
+        room_identity
+        |> Map.put(:name, room.name)
+        |> Messages.RoomInvite.new()
+        |> Dialogs.add_new_message(me, dialog)
+        |> RoomInviteIndex.add(dialog, me)
+      end
+    end)
+
     noreply(socket)
   end
 
@@ -624,6 +659,6 @@ defmodule ChatWeb.MainLive.Index do
     |> Page.Dialog.init()
     |> Page.Logout.init()
     |> Page.Shared.track_onliners_presence()
-    |> Page.RoomRouter.route_live_action()
+    |> Page.LiveRouter.action()
   end
 end