+++ lib/chat_web/live/main_live/page/admin_panel.ex
@@ -1,20 +1,27 @@
 defmodule ChatWeb.MainLive.Page.AdminPanel do
   @moduledoc "Admin functions page"
   import Phoenix.Component, only: [assign: 3]
-  import Phoenix.LiveView, only: [push_event: 3, send_update: 2]
+  import Phoenix.LiveView, only: [push_event: 3, put_flash: 3, send_update: 2]
+  import ChatWeb.LiveHelpers, only: [open_modal: 2, close_modal: 1]
 
   alias Chat.RoomInviteIndex
   alias Phoenix.PubSub
 
   alias Chat.AdminRoom
+  alias Chat.ChunkedFiles
+  alias Chat.ChunkedFilesMultisecret
   alias Chat.Db.{FreeSpacesPoller, FreeSpacesSupervisor}
   alias Chat.Dialogs
+  alias Chat.FileIndex
+  alias Chat.MemoIndex
   alias Chat.Messages
   alias Chat.Rooms
   alias Chat.Rooms.RoomsBroker
+  alias Chat.Upload.UploadKey
   alias Chat.User
   alias Chat.User.UsersBroker
   alias ChatWeb.MainLive.Admin.CargoWeightSensorForm
+  alias ChatWeb.MainLive.Admin.FirmwareUpgradeForm
   alias ChatWeb.Router.Helpers, as: Routes
 
   @admin_topic "chat::admin"
@@ -208,16 +215,66 @@ defmodule ChatWeb.MainLive.Page.AdminPanel do
     socket
   end
 
-  def create_cargo_user(socket, name) do
+  def create_cargo_user(socket, {identity, backup_content, backup_entry}) do
     cargo_user =
-      name
-      |> User.login()
+      identity
       |> tap(&AdminRoom.store_cargo_user/1)
       |> tap(&User.register/1)
       |> tap(&UsersBroker.put/1)
 
+    AdminRoom.admin_list()
+    |> Enum.each(fn admin_card ->
+      dialog = %{b_key: b_key} = Dialogs.open(cargo_user, admin_card)
+      destination = %{dialog: dialog, pub_key: Base.encode16(b_key, case: :lower), type: :dialog}
+      file_key = UploadKey.new(destination, cargo_user.public_key, backup_entry)
+      file_secret = ChunkedFiles.new_upload(file_key)
+
+      :ok = save_file({file_key, backup_content}, {backup_entry.client_size, file_secret})
+
+      now = DateTime.utc_now() |> DateTime.to_unix()
+
+      text =
+        "The backup `#{backup_entry.client_name}` is not encrypted. Do not share it with anyone."
+
+      %Messages.Text{text: text, timestamp: now}
+      |> Dialogs.add_new_message(identity, dialog)
+      |> MemoIndex.add(dialog, identity)
+
+      {_index, msg} =
+        backup_entry
+        |> Messages.File.new(file_key, file_secret, now)
+        |> Dialogs.add_new_message(cargo_user, dialog)
+
+      FileIndex.save(file_key, dialog.a_key, msg.id, file_secret)
+      FileIndex.save(file_key, dialog.b_key, msg.id, file_secret)
+    end)
+
     socket
     |> assign(:cargo_user, cargo_user)
+  end
+
+  def upgrade_firmware_confirmation(socket) do
+    socket
+    |> open_modal(ChatWeb.MainLive.Modals.ConfirmFirmwareUpgrade)
+  end
+
+  def upgrade_firmware(socket) do
+    send_update(FirmwareUpgradeForm, id: :firmware_upgrade_form, step: :upgrade)
+
+    socket
+    |> close_modal()
+  end
+
+  def notify_firmware_upgraded(socket) do
+    send_update(FirmwareUpgradeForm, id: :firmware_upgrade_form, substep: :done)
+
+    socket
+    |> put_flash(:info, "Firmware upgraded")
+  end
+
+  defp save_file({file_key, content}, {size, file_secret}) do
+    ChunkedFilesMultisecret.generate(file_key, size, file_secret)
+    ChunkedFiles.save_upload_chunk(file_key, {0, max(size - 1, 0)}, size, content)
   end
 
   defp request_platform(message),